// github/types/paths/user/starred/{owner}/{repo}.types.ts
export type activityCheckRepoIsStarredByAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityCheckRepoIsStarredByAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: basic_error;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Note that you'll need to set `Content-Length` to zero when calling out to this endpoint. For more information, see "[HTTP method](https://docs.github.com/rest/guides/getting-started-with-the-rest-api#http-method)."
 */
export type activityStarRepoForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityStarRepoForAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Unstar a repository that the authenticated user has previously starred.
 */
export type activityUnstarRepoForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: activityUnstarRepoForAuthenticatedUser_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type activityCheckRepoIsStarredByAuthenticatedUser_Path = {
  owner: string;
  repo: string;
};

export type activityStarRepoForAuthenticatedUser_Path = {
  owner: string;
  repo: string;
};

export type activityUnstarRepoForAuthenticatedUser_Path = {
  owner: string;
  repo: string;
};